backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1
pgvector==0.3.6

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# LLM
groq==0.15.0
langchain==0.3.14
langchain-groq==0.2.4
langgraph==0.2.62

# Task queue
celery[redis]==5.4.0
redis==5.2.1

# PDF processing
PyPDF2==3.0.1
reportlab==4.2.5
python-docx==1.1.2

# Browser automation
playwright==1.49.1

# Email
aiosmtplib==3.0.2

# Embeddings (local, free)
sentence-transformers==3.3.1

# Monitoring
prometheus-client==0.21.0

# Utilities
pydantic[email]==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
python-dotenv==1.0.1
